TIP-842: Move attribute group normalizer to Akeneo\Pim\Structure
