+++ src/Akeneo/Tool/Bundle/ApiBundle/Resources/config/normalizers.yml
@@ -3,8 +3,7 @@ parameters:
     pim_api.normalizer.exception.violation.class: Akeneo\Tool\Component\Api\Normalizer\Exception\ViolationNormalizer
     pim_api.normalizer.collection.class: Akeneo\Tool\Component\Api\Normalizer\CollectionNormalizer
     pim_api.normalizer.product.class: Akeneo\Tool\Component\Api\Normalizer\ProductNormalizer
-    pim_api.normalizer.category.class: Akeneo\Tool\Component\Api\Normalizer\CategoryNormalizer
-    pim_api.normalizer.attribute_group.class: Akeneo\Tool\Component\Api\Normalizer\AttributeGroupNormalizer
+    pim_api.normalizer.attribute_group.class: Akeneo\Pim\Structure\Component\Normalizer\ExternalApi\AttributeGroupNormalizer
     pim_api.normalizer.file.class: Akeneo\Tool\Component\Api\Normalizer\FileNormalizer
     pim_api.normalizer.association_type.class: Akeneo\Tool\Component\Api\Normalizer\AssociationTypeNormalizer
 
@@ -46,12 +45,6 @@ services:
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
 
-    pim_api.normalizer.category:
-        class: '%pim_api.normalizer.category.class%'
-        arguments:
-            - '@pim_catalog.normalizer.standard.category'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
 
     pim_api.normalizer.attribute_group:
         class: '%pim_api.normalizer.attribute_group.class%'
